Adding the first real post
Added a post about why I created the blog. It's in a new category jsut for posts about the creation of the dev blog. I intend to put some more in there soon about the construction process, and then one about the search function if i ever end up creating that.